services/requirements.txt
confluent-kafka
faker
psycopg2-binary==2.9.9